fix: refactor kebab service and added implementation of that
fix: version 21 -> 23
